+++ SystemML/SystemML/DML/src/main/java/com/ibm/bi/dml/runtime/instructions/mr/ReplicateInstruction.java
@@ -104,12 +104,12 @@ public class ReplicateInstruction extends UnaryMRInstructionBase
 						//(e.g., M is Nx2700, blocksize=1000 -> numRep 2 because original block passed to index 1)
 						if( blockColFactor<=1 ) //blocksize should be 1000 or similar
 							LOG.warn("Block size of input matrix is: brlen="+blockRowFactor+", bclen="+blockColFactor+".");
-						int numRep = (int)Math.ceil((double)_lenM / blockColFactor) - 1; 
+						long numRep = (int)Math.ceil((double)_lenM / blockColFactor) - 1; 
 						
 						//replicate block (number of replicates is potentially unbounded, however,
 						//because the vector is not modified we can passed the original data and
 						//hence the memory overhead is very small)
-						for( int i=0; i<numRep; i++ ){
+						for( long i=0; i<numRep; i++ ){
 							IndexedMatrixValue repV = cachedValues.holdPlace(output, valueClass);
 							MatrixIndexes repIX= repV.getIndexes();
 							repIX.setIndexes(inIx.getRowIndex(), 2+i);
@@ -133,12 +133,12 @@ public class ReplicateInstruction extends UnaryMRInstructionBase
 						//(e.g., M is Nx2700, blocksize=1000 -> numRep 2 because original block passed to index 1)
 						if( blockRowFactor<=1 ) //blocksize should be 1000 or similar
 							LOG.warn("Block size of input matrix is: brlen="+blockRowFactor+", bclen="+blockColFactor+".");
-						int numRep = (int)Math.ceil((double)_lenM / blockRowFactor) - 1; 
+						long numRep = (int)Math.ceil((double)_lenM / blockRowFactor) - 1; 
 						
 						//replicate block (number of replicates is potentially unbounded, however,
 						//because the vector is not modified we can passed the original data and
 						//hence the memory overhead is very small)
-						for( int i=0; i<numRep; i++ ){
+						for( long i=0; i<numRep; i++ ){
 							IndexedMatrixValue repV = cachedValues.holdPlace(output, valueClass);
 							MatrixIndexes repIX= repV.getIndexes();
 							repIX.setIndexes(2+i, inIx.getColumnIndex());
